[Fix]Overflow when setting all bits of a byte
Calculate as u16, then convert to u8.

+++ src/lib.rs
@@ -19,7 +19,8 @@
 /// ```
 pub fn set(start_byte: usize, start_bit: usize, num_of_bits: usize) -> () {
     unsafe {
-        *(start_byte as *mut u8) = (1 << (start_bit + num_of_bits)) - (1 << start_bit);
+        *(start_byte as *mut u8) =
+            ((1 << (start_bit + num_of_bits)) as u16 - (1 << start_bit) as u16) as u8;
     };
 }
